Refactor: Note 리팩토링
