+++ bovo_client/src/components/note/Note.module.css
