+++ bovo_client/src/mocks/dummy.json
@@ -8,7 +8,7 @@
             "author": "저자가 엄청 길어지면 어떻게 처리하는지 예시 - 저자가 엄청 길어지면 어떻게 처리하는지 예시",
             "start_date": "11. 22. 33",
             "end_date": "99,88,77",
-            "star": 10,
+            "star": 5,
             "memos":[
                 {
                     "memo_id": 1234,
